WORK IN PROGRESS - Fix references in timeline triggers; add scrubber; migrate complex packagers
SimpleEventTrigges and ComplexEventTrigger playables now dynamically
repopulate themselves as has been implemented in the rest of the codebase.

I added a first pass at the NavigationController, where the scrubber
functionality is housed. (Most of the logic for this was already contained
inside of MasterSequence).

I also revised ComplexEventConfigurableTriggers to use references rather
than normal values. Just realized I need to also change the code so
they trigger complex events with the correct values now - will do that
for the next commit.

Note: While working with the timeline triggers, I caused all assets in the
project to reserialize, so many of the committed files here are due to
meta files being updated.

Next steps are to finalize the ComplexEventConfigurableTrigger, do a run
through of the app to make sure things work, wrap all code that needs to be
wrapped the with #if Unity_Editor conditional, deploy to device, then
run through all of our active projects to make sure everything gets migrated
/ imported correctly, cleaning up bugs and removing any code that's
no longer in use.