WORK IN PROGRESS - Add 'AND' functionality to ConditionResponse
Realized that it would be fairly easy to add the
'AND' functionality to our ConditionResponse logic.
Didn't take long to implement but I think will do a lot to make
creating logic within the editor much easier.

Also, fixed the ComplexEventConfigurableTrigger so that it
raises events correctly.

I haven't done a run-through of any apps yet, so not sure
how well everything works in Play mode yet.

Next steps are to begin thorough testing and cleaning up code.