Revise references to AppSettings throughout project; successfully deploy to device
At long last, dealt with compiler errors in deploying Carrier Pigeon 19
to device and resolved all issues. Namely, this involved ensuring that
editor only blocks were marked correctly, and making sure all references
were populated and serialized properly. Turns out that the way I was
referencing AppSettings throughout the project was not working well,
so I introduced a new AppSettingsReference script, which mirrors the
way other variable references work.

The frame rate for Riveted is atrocious, not sure if due to sheer number
of game objects or because I tried to make the GradientShift work with
signal reference updates as opposed to working with the Update() function.
Would like to take a quick look at this and think through performance
optimization as I begin testing with Legend and Violino.