Fix bugs related to sequencing refactor; convert certain AppSettings values to static
Previous optimizations related to the TimelineUtilsBehaviour
were causing some breakage. Namely, having one source of truth
for accessing currentTime prevented us from having multiple
sequences / timelines from executing at once. To fix this, I moved
the currentTime value to TrackAssetConfig, which means that
each sequence / timeline as a whole now has its own independent currentTime.
(I revised mixer behaviours and playables in order to reflect this
change).

Realized also that holding a direct reference to an instance of
AppSettings in persistent editor variables was breaking version control,
so I converted some values in AppSettings to static. These are values
that should never be deployed to production (logEventCallersAndListeners,
for example).

In the event that we need to debug within a build, the easiest thing to do
is just flip the bool needed in code, then switch it back once debugging
is complete.

Next steps are to test all this new functionality, including bug fixes,
with Violino. After fixing any bugs there, next steps are to test loading
bar / spinner, and create a dynamic camera based on current scene width / height
for the AppUtils UI.