Add utility properties to ActionTriggerBehaviour; rename assets
Add relatively small changes related to quality-of-life issues
with the ActionTriggerBehaviour. We now have a ManualOverride
bool to manually edit serializedPropertyPath if the serialization
gets out of whack. Also fix bugs related to parsing UnityEvents
and the AudioForwardReverseTrack.

Lastly, deleted some Fader persistent variables that are no longer
needed, and moved assets so they are easier to find.

Next steps are to remove all assets that are no longer needed,
which will conclude the broad strokes of the AppSettings refactor.
Still need to polish and make sure there are no bugs when it comes
to creating default settings, but otherwise everything seems to be
working well overall.

After that, will move onto testing the loading bar / spinner on
iOS, followed by creating the dynamic scene camera for AppUtils,
polishign up scrubbing, implementing a bookmarking module, and then
refactoring the autorun functionality.