Delete unused assets and obsolete scripts
As part of the AppSettings refactor and moving towards
the beta version of Maestro, deleted all the assets that
are now created dynamically underneath of AppSettings and
related settings scripts. Also took the opportunity to go
through the codebase and remove any scripts that (I believe)
are no longer in use.

Next steps are to test all these revisions with Riveted and
Legend to ensure everything is still working properly, then
move onto testing the loading bar / spinner and creating the
dynamic scene camera for AppUtils.