WORK IN PROGRESS - Revise RegisterDependencies, is working w/o special cases
The RegisterDependencies tool is working now for basic cases. I still
need to add the event description data for ActionTriggers and ConditionResponses,
as well as do a code cleanup. Also need to revise the playables registration
to mirror the formatting of the component registration.

IMPORTANT: I revised IntVariable to allow for easy debugging; this should
be removed when the RegisterDependencies tool is complete.