Continue working on RegisterDependencies tool; fix overwriting bugs
Each registration within the RegisterDependencies tool was overwriting
the previous ones - turns out it was because we were actually overwriting
the file each time. To fix this, I reintroduced the WriteDat() function
into every applicable method so that we can build up registration files
as we parse the dependencies.

This isn't clean - ideally we would get a collection of all of our data,
merge the objects, then write out the final files. But it does seem to be
working, with the bonus that we don't need to use any hefty JSON libraries
- the SimpleJSON utility class we're currently using works just fine. I'm
going to say then that the RegisterDependencies tool is done for now, and
any cleanup here will be a nice to have.

Last steps are to remove errant code, including debug code, make the tool
works in all scenes, then move onto polishing scrubbing, creating a dynamic
app utils camera, and testing the loading bar / spinner.