Fix loading bar, fix spinner functionality, Tutorial responsive element fixes
Made the loading bar and spinner functionality robust. The fader
now communicates with the loading bar using callbacks to make sure
we can sequence fading, the appearance of the loading bar, fading
out the loading bar, and fading a scene back in properly.

Also discovered that the spinner has been working all along, just that
we couldn't see it due to having a black overlay; changing the activity
indicator style to white fixed this.

The responsive elements in Tutorial related to ScrollSnapController
were not getting updated at the right time, so changed their priority to
fix this.

Next steps are to create a dynamic AppUtils camera, followed by fixing a
bug related to swiping at branching paths, polishing scrubbing, and adding
bookmarking functionality.