Create Current Aspect Ratio Camera; modify prefabs; fix bug in Master Sequence
Created a script to dynamically resize the viewport of certain cameras,
namely the RequestAppUtils and AppUtils cameras, so that they correspond
to the current active story scene.

Also, updated prefabs as a result, introducing a new base camera
prefa, of which the Current Aspect Ratio Camera and Responsive Camera
are both children.

The Master Sequence wasn't configuring its data due to a modification
made during the progress bar / spinner optimization commits, so I added
that functionality back in.

Next steps are to work on polishing the scrubbing functionality, as well
as implementing the bookmarking functionality.