Begin polishing scrubbing issues; add playable compatibility with scrubber
Sequencing can break when using the scrubber, so began adding
extra checks and logic to allow not only for the scrubber, but
bookmarking as well.

Namely, the logic within the TimelineTriggers has changed to be
more robust, ensuring that events are fired correctly when moving
forward or backwards in time.

Also made sure that LerpToTarger playables are resetting their
bindings only when the scrubber is actively being pressed, and not
simply when the AppUtils get summoned.

Lastly, added flags to allow FadePlay audio tracks to also be used
to initialize audio on sequences adjacent to where an originating
FadePlay track is placed.