Add Bookmarker; continue polishing scrubber; add ArrowIndicator; modify DynamicLayoutController
Principal changes in this commit include adding the data
necessary to make the Bookmarker work. Additional changes relate
to polishing for the scrubber, as well as prefabs
and logic to handle the ArrowIndicator.

The Bookmarker works by using several new fields in SystemSettings,
namely HasBookmark, LastOpenedScene, LastLoadedSequence, and
LastLoadedSequenceTime. It works in conjunction with these variables
and a RootConfig to restore a user's position in a story when
reopening the app.

Whenever the app is closed from a scene containing a Bookmarker,
it will set the HasBookmark boolean to true and save the scene,
sequence, and time using the UserDataController.

Regarding loading the bookmark, I modified the flow of
story scene startup to accommodate the new functionality.
Whenever a scene with a Root Config is opened, the RootConfig
calls an event, CongigurationCompleted. to trigger the Bookmarker,
which will activate the saved sequence at the saved time
using a MasterSequence's SetToElapsedTime() method.
This will override default sequence settings. Once this is
complete, the Bookmarker calls the event BookmarkLoadingCompleted,
which will signal to a PrepareScene ActionTriggerBehaviour
that it can reset logic and trigger a fade in.

Importantly, TimelineTriggers listen to BookmarkLoadingCompleted
so we can ensure that sound effects and other events don't get
triggered during the bookmark execution unless we want them to.
(I also added some bools to determine whether these events should
be executed when AppUtils are requested as well).

With all this in mind, I revised the UserDataController for robustness
to handle saving more than one scriptable object value at once.
Also introduced a complex event, DeleteStoredData, to clear the
player preferences if desired. I also revised the RootConfig,
MasterSequence, and SequenceConfig to not reset sequences or
calculate data on Awake() anymore so as not to interfere
with Bookmarker.

Miscellaneous fixes include revising touchable to not execute
when beneath UI.

The ArrowIndicator is fairly straightforward. Via a NavigationController
raising a RefreshAppUtils complex event, it listens to the ySwipeAxis
and xSwipeAxis, then activates / deactivates game objects containing
the relevant images accordingly.

Next steps are to introduce EnableScrubber and EnableArrowIndicator
bools, then move onto refactoring the Autorun functionality.