Add ability to enable / disable arrow indicator and scrubber
Added arrowIndicatorEnabled and scrubberEnabled bools to the
input group settings.

These values should get reset as part of the game state.
While testing, I realized that the ActionTriggerBehaviour
was resetting the game state at the wrong time, so that has
been fixed.

Added a condition response ActionTriggerBehaviour to the
Menus prefab to enable / disable the arrow indicator and
scrubber based on their related bools.

Also realized that the triggerOnStart bool inside of
ConditionResponse hadn't been cleaned ups, so went ahead
and removed that along with related code that's no longer
being used.