WORK IN PROGRESS - Begin refactoring Autorun; rename classes; consolidate functionality
Big changes here include consolidating all of the functionality related
to updating sequences into the new SequenceController class. Also, rather
than manually advance and evaluate a sequence when autoplaying, we're
now using the PlayableDirector's built-in Play() method, and manipulating
the speed of the playable graph depending on the behaviour we want.

Importantly, the Play() method only works when playing a timeline while
moving *forward*, but not while moving backward, so we still need to evaluate
the timeline manually when reversing.

The reason we're using the Play() functionality now is that, unfortunately,
Timeline doesn't evaluate native Audio or Animation tracks when updating a
playable graph manually... however, I've found a way around this. If you call
Play() on a playable director, then set the speed to 0, you can still scrub the
timeline as we've been doing, and the animation tracks work. In addition, if you
set the speed back to 1 at any time, the audio will continue evaluating at the
correct spot.

The only issue I've run into so far with this is that audio tracks, when the speed
is set to 0 and the timeline is updated manually, will
emit clicks and pops, so I've written some functionality to set audio
source volumes to 0 while we manually update the timeline. IMPORTANT: We can
only set the volume to 0 for audio tracks that contain an audio source component
as a binding (there may be a way to also mute ones w/o a binding, a search
on the Unity forums suggested that would be difficult or impossible though).

Ultimately, being able to use Unity's native audio tracks now is *very* nice,
as it will allow us to create animations that depend on precise execution of audio,
and muting while manually updating or reversing during these sequences
seems to make intuitive sense Also, note that we still have the ability to play
audio independently of the timeline using FadePlay tracks or complex events,
 in which case the audio will continue playing during manual updates or reversing.

Also important to note is that we still need to be able to pause a director
that's playing forward depending on the sequence configuration. In order to
make this work, I've introduced an event, TimelineUpdated, into what was originally
called TrackAssetConfig, and hooked up the TimelineUtilsMixerBehaviour to trigger
this event on PrepareFrame(). What this means is that we get a callback for every
frame when the playable director is playing, which allows us to update a sequence
with the timeline's current time using a corresponding SequenceUpdated event
in SequenceController, which the AutorunController can then listen to in order
to determine whether the newly evaluated time falls within the autorun extents.
(Unfortunately, Unity doesn't provide a built-in way for listening to when
a playable graph has been updated - which is why I've coded these events).

Other important points: Due to the way the updating / callback functionality
works using the Play() method when going forward, we now need to keep track
of which AutorunExtents is currently active in order to be able to pause
at the correct moment. There's a lot of logic inside of RefreshAutoplay
within the Autoplayer to make this work (this still needs to be cleaned up).

That covers the broad strokes of how the Autorun functionality should work now.
TrackAssetConfig has been renamed to TimelineInstanceConfig to better reflect
its functionality, and next steps are for me to continue cleaning this up,
removing old files, and consolidating scripts so that all of this
is easier to follow within the code.