WORK IN PROGRESS - Autoplay refactor code cleanup
Consolidated a lot of scripts and added comments related to the
functionality described in the previous commit.

SequenceController has seen several methods remove or revised,
so this has potential to break the Joiner and MasterSequence.
MasterSequence's SetToElapsedTime method has also been revised,
and the checks for forward and backward autoplaying have been
merged into the RefreshAutoplayStatus method in the Autoplayer.

Next steps are to test with other projects to ensure all is working
properly, from the scrubber, to bookmarking, to swiping, momentum,
and autoplaying, as well as refactor the Lerper to correspond
to the refactor.