Hook Lerper up to new Autorun refactor; fix related bugs
The Lerper now works similarly to the Autoplayer and uses
Unity's built-in Play() method to autoplay sequences going forward.

While implementing this, ran into several small issues related
to the refactor, namely starting / pausing autoplay correctly
between sequences. Previously, all we needed to do was switch
on an autoplay boolean, and since the autoplayer ran all of its
logic through the Update() method, it would automatically start
playing as soon as a sequence entered an interval.

However, since the autoplayer is now dependent on the SequenceUpdated
callback, I added a method, ActivateAutoplayEnabledAndRefresh, which
forces a reevaluation of the autoplay status after being enabled.

Most of the internal functions related to the Autoplayer won't need
to make use of this, however it seems to be a useful method
for interacting with the autoplayer via an ActionTriggerBehaviour.

Made other small modifications to make sure the bookmarker and
autoplayer don't interfere with one another, and to make sure
that behaviour works as expected when at the extremes (beginning
and end) of a sequence.

Next step is to test all of this in Riveted, then move onto
refactoring the video player.