Revise Video Player functionality; add 'Pause' functionality to SequenceController
I cleaned up the previous implementation of our video player.
Timeline-driven videos are now controlled via a ManualVideoPlayer
script, which controls a ManualVideoPlayerItem component, which
contains logic for using a custom color variable built in.
Note that the ManualVideoPlayer is intended to only play
videos moving forward in time, NOT backward, and is not
meant to work in conjunction with the Sequencing functionality.

That being said, the ForwardReverseVideoPlayer builds on the ManualVideoPlayer
by providing handling for reversing videos in Android, thus
allowing us to combine it with our Sequencing logic.  However,
despite my best efforts, this doesn't work great, as videos generally
aren't made to be scrubbed when being viewed, so we'll likely
just use the ManualVideoPlayer in most cases and won't give
users the option of swiping through videos.

Lastly, I created a SimpleVideoPlayer prefab that combines several
other components to create a less robust video implementation
if we're looking for something more traditional that's not driven
by Timeline. Additional items to potentially be added to the
SimpleVideoPlayer include the ability to scrub, adjust volume, etc.,
though can tackle this when / if we end up needing it. Note: I envision
the SimpleVideoPlayer being similar to a video player you'd find
anywhere on the internet, with controls, seeking, etc.

I haven't tested all of this extensively on device, namely the
SimpleVideoPlayer, though the ManualVideoPlayer and ForwardReverseVideoPlayer
seem to be working well enough on Android, so can make any small
revisions to this the next time I deploy.

*Also, note that I added a couple of new delegates to allow for pausing of
sequences via the TimelineAssetConfig, which is used by the ForwardReverseVideoPlayer
to briefly pause sequencing logic in order to let the device catch up
before continuing with video evaluation*

Next steps then, are to move onto implementing Unity analytics
and dealing with GDPR.