WORK IN PROGRESS - Clean up Maestro Tools; perform AudioLerpSnapshot refactor
Lots of Maestro changes, adding editor tools and cleaning up
code along the way.

Made in particular some changes to the VideoPlayer prefabs so
that they appear without errors when being created for the first time.

Added several prefabs to the ModuleReferences scriptable object so
they can be instantiated from the control panel.

IMPORTANT: I introduced a breaking change with AudioLerpSnapshot.
In order to simplify the workflow here, I am now requiring that
an Off and On snapshot be specified, in that order, for crossfading
to work. (I still need to write the rest of the functionality
to allow us to create AudioLerpSnapshotClips easily)

Next steps are to commit changes to Legend, which makes extensive
use of the AudioLerpSnapshot playable, and then finish up the audio
editor tools.

Also important to note: Unity currently does not allow for the creation
of audio mixers via scripting, so to get around this, I created an
AudioMixerTemplate, which we then duplicate using the Edit/Duplicate
command, and then migrate to the correct folder accordingly. (We need
to make use of a mixer that's been created in a particular way,
namely with Off and On snapshots, in order for crossfading to work).