Add support for horizontal orientations
To make for a better editing experience when working
primarily horizontally, I added some scaling equationa and
cleaned up the aspect ratio code in DynamicLayoutController.

Interestingly, a linear equation works just fine for the
calculated horizontal responsive width, whereas we needed
an exponential one for the vertical orientation.

The new equation can be found in Utils.