Fix input module priority implementation; add horizontal support; quality-of-life sequencing fixes
Big changes include fixing the autoplayer input module so that it
correctly ends its input after completing an interval. Along with this,
created appropriate events and logic to halt the momentum applier
whenever autoplay begins, and added a new OnTouchStart event to allow
for granular control of how an autoplayer can be overriden.

Now, you can adjust the priority and/or events of the sequencing modules
to create the kind of interactions you need for a particular story.

In addition, cleaned up the implementation of the DynamicLayoutController's
horizontal support. In short, we can now design stories in the horizontal
format, same as the vertical format, with dynamic width-height containers.
The only thing that's missing is creating a robust, standard implementation
of menus and app utils for horizontal orientations.

Lastly, added some small quality-of-life fixes related to the RootConfig to
ensure that we can easily play through a story from either the story's scene
itself or the initializer without needing to manually create the appropriate
conditions so that everything works each time we enter play mode.

NOTE: Due to the new functionality with the DynamicLayoutController, I've altered
the callback chain at the beginning of story scenes so that all scene preparation
continues to happen smoothly. The standard events for a scene setup, then, is as follows:

-Scene Controller triggers a scene load, calls SingleSceneLoadCompleted
-RootConfig performs configuration, calls ConfigurationCompleted
-Bookmarker performs bookmark loading, calls BookmarkLoadingCompleted
-DynamicLayoutController renders scene layout, calls DynamicLayoutRenderCompleted
-PrepareSceneActionTrigger calls TriggerFadeIn; scene can begin
-Fader calls FadeInCompleted
-Optionally, an additional ActionTrigger can call AutoplayActivate
to start sequence animations