Add support for mouse; add platform logic to action trigger; fix manual autoplayer
Big change in this commit is adding further support for desktop
via the MouseScrollMonitor.

I also added a GenericInputController, which allows for modifying of
sequences with any arbitrary float value. For now, this is how the
MouseScrollMonitor interacts with the RootConfig – I'm not sure it
needs to be coded if everything can be accomplished in the editor.

I added Platform Logic for iOS, Android, Standalone, and WebGL, in order
to make handling for specific scenarios more robust in the ActionTrigger.

Lastly, the autoplayer and manual autoplayer were not playing nicely together,
so I fixed that. Namely, the manual autoplayer's functionality is slightly
different in that it needs to register itself when it triggers autoplay, and
only then, whereas the regular autoplayer refreshes its status every
time the sequence is modified.

To solve this, I created a new child ManualAutoplayer script, overrode some
methods, and updated the integration tests to reflect these changes. The
implementation seems to be working for now.